Added APIClient class and replaced categories api functions
