+++ CakeCrafter.Frontend/src/pages/CakesCategories/CakesCategories.jsx
@@ -3,45 +3,6 @@ import DisplayCategory from './Components/DisplayCategory';
 import { Input, Button, Menu, MenuButton, MenuList, MenuItem, Box } from '@chakra-ui/react';
 import { AddIcon } from '@chakra-ui/icons';
 
-export async function GetCategories(){
-
-      const response = await fetch("http://localhost:5000/api/categories");
-      const jsonResponse = await response.json();
-
-    return jsonResponse;
-}
-
-export async function CreateCategory({ request }){
-    const formData = await request.formData();
-    const newCategory = formData.get("NewCategory");
-    const response = await fetch("http://localhost:5000/api/categories",
-                                {
-                                    method: 'POST',
-                                    headers: { "Content-Type": "application/json" },
-                                    body: JSON.stringify({name: newCategory}),
-                                });
-    return redirect('/categories');
-}
-
-export async function UpdateCategory({ params, request }){
-    const formData = await request.formData();
-    const newCategoryName = formData.get("CategoryName");
-    const response = await fetch(`http://localhost:5000/api/categories/${params.categoryId}`,
-                                {
-                                    method: 'PUT',
-                                    headers: { "Content-Type": "application/json" },
-                                    body: JSON.stringify({id: params.categoryId, name: newCategoryName}),
-                                });
-    return redirect('/categories');
-}
-
-export async function DeleteCategory({ params }){
-    const response = await fetch(`http://localhost:5000/api/categories/${params.categoryId}`,
-                                {
-                                    method: 'DELETE',
-                                });
-    return redirect('/categories');
-}
 
 
 function CakesCategories() {
